requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
qdrant-client>=1.9.0
arxiv>=2.0.0
fastmcp>=0.2.0
langfuse==2.42.0
pydantic>=2.5.0
dill>=0.3.8
json-repair>=0.2.0
tqdm>=4.66.0
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-qdrant>=0.2.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
httpx>=0.25.0
fastapi
uvicorn